moved equipment component out of client component
