+++ atecAPP-web/src/main/webapp/src/app/equipment/equipment.component.html
@@ -25,10 +25,10 @@
     <textarea class="input-label"
               style="width: 100%" rows="1"
               [(ngModel)]="equipment.type"
-              (change)="updateEquipment(equipment)"></textarea>
+              (change)="updateEquipment()"></textarea>
     <textarea class="input-label"
               style="width: 100%" rows="1"
               [(ngModel)]="equipment.code"
-              (change)="updateEquipment(equipment)"></textarea>
+              (change)="updateEquipment()"></textarea>
   </div>
 </div>
