+++ atecAPP-web/src/main/webapp/src/app/equipment/equipment.component.ts
@@ -1,5 +1,6 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {Equipment} from "../model/equipment";
+import {Equipment} from "./model/equipment";
+import {ClientService} from "../service/client.service";
 
 @Component({
   selector: 'app-equipment',
@@ -12,11 +13,26 @@ export class EquipmentComponent implements OnInit {
   @Output() equipmentUpdated: EventEmitter<Equipment> = new EventEmitter<Equipment>();
   @Output() equipmentDeleted: EventEmitter<Equipment> = new EventEmitter<Equipment>();
 
-  constructor() { }
+  equipmentList: Equipment[]=[];
+
+  constructor(private clientService: ClientService) { }
 
   ngOnInit() {
+    this.getAllEquipment();
   }
 
+  getAllEquipment() {
+    this.clientService.getAllEquipment().subscribe(
+      response => {
+
+        this.equipmentList = response;
+
+      },
+      error => {
+        alert("An error has occurred while getting equipment list!")
+      }
+    )
+  }
   updateEquipment(){
 
     this.equipmentUpdated.emit(this.equipment);
