+++ atecAPP-web/src/main/webapp/src/app/equipment/model/equipment.spec.ts
